Init
Added file(s) to the repository. I'm going to be trying to break up the script into smaller modules, and add more functionality, but for now, it's one big script with everything lumped into it, and it doesn't quite work as intended...this is more a public place for the code when I ask for help.